Nightly Progress for July 21
I've decided to abandon the Jtree for now. I don't think I'll be able to get it to work in time, and there aren't very many good tutorials out on it except for the Javadocs. Trying to adapt it to my use case will be hard at my skill level, and I think a drop-down menu and list will suffice (and probably look cleaner).