More success with lists
I am having more success trying to implement a list, but I am noticing that perhaps a table would be cleaner. To prevent constant hopping between ideas without producing a functional item, I will continue on this path and bring it to completion. Afterwards, I can pull a branch and try to "clean it up".