An (almost) functioning program
The "Admin Screen" works, with the drop down menus functioning and population properly, as well as the part lists. I had a data structure worked out using folders and CSV files. Each sub folder contained multiple CSVs, which represented the Arraylist of the different classes deeper down the hierarchy. I'm rather happy with the result.

Currently, the delete button doesn't work. It isn't deleting parts properly, and I do not know why. I still need to implement the Edit button, but as of right now, these aren't "Critical" to the functioning of the app.

The Admin screen layout could do with a few more buttons to delete the relevant Engines and assemblies. That might be complex, though, and not feasible under the given time constraints.